Update remote_control.dart (#14)
compute scaledVolume as integer to fix #13 
(not tested)

+++ lib/remote_control.dart
@@ -435,7 +435,7 @@ class _RemoteControlState extends State<RemoteControl> {
   }
   
   _volumePercent(int percent) async {
-    var scaledVolume = percent * VolumeSliderScaleFactor;
+    var scaledVolume = (percent * VolumeSliderScaleFactor).toInt();
     var response = await _statusRequest({
       'command': 'volume',
       'val': '$scaledVolume',
